+++ src/components/signup.jsx
@@ -1,27 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
-import { green } from '@material-ui/core/colors';
-import CircularProgress from '@material-ui/core/CircularProgress';
 import InputField from './common/inputField';
 import '../assets/styles/signup.scss';
 import { NavLink } from 'react-router-dom';
 import SignupButton from './common/signupButton';
 import FormError from './common/formError';
+import Spinner from './common/spinner';
 
-const useStyles = makeStyles(theme => ({
-  progressdiv: {
-    display: 'flex',
-    justifyContent: 'center'
-  },
-  progress: {
-    margin: theme.spacing(2),
-    color: green[500],
-  },
-}));
 
 const Signup = ({ handleChange, handleSubmit, isLoading, error, errorMessage }) => {
-  const classes = useStyles();
   return (
     <div className="signupcontainer">
       <div className="signupcontainer__header">
@@ -36,11 +23,7 @@ const Signup = ({ handleChange, handleSubmit, isLoading, error, errorMessage }) 
           error && <FormError message={errorMessage}/>
         }
         {!isLoading && <SignupButton form="signup" handleSubmit={handleSubmit}/>}
-        {isLoading && (
-          <div className={classes.progressdiv}>
-            <CircularProgress className={classes.progress} />
-          </div>
-        )}
+        {isLoading && <Spinner/>}
         <div className="container__form__linksignup">
           <p>Have account already?<NavLink to="/login">Login</NavLink></p>
         </div>
