+++ src/containers/signup.jsx
@@ -15,6 +15,11 @@ class SignupView extends Component {
     errorMessage: '',
   }
 
+  /**
+   * handles user input
+   * @param event
+   * @returns {void}
+   */
   handleChange = (event) => {
     event.preventDefault();
     const { name, value } = event.target;
@@ -25,10 +30,22 @@ class SignupView extends Component {
     });
   }
 
+  /**
+   * toogles boolean value in state property
+   * @param name
+   * @param value
+   * @returns {void}
+   */
   toggleState = (name, value) => {
     this.setState({ [name]: !value });
   }
 
+  /**
+   * handles submission of information to server
+   * triggers the middleware function to creates the action
+   * @param event
+   * @returns {void}
+   */
   handleSubmit = async (event) => {
     this.toggleState('isLoading', this.state.isLoading);
     event.preventDefault();
